api_server-network/inference-server/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.7.1
joblib==1.4.2
scikit-learn==1.4.2
python-dotenv==1.0.1
mlflow
boto3
statsmodels
pandas==2.2.3
prophet==1.1.7
numpy==2.2.4
psycopg2-binary